requirements.txt
fastapi
httpx
uvicorn
locust
python-dotenv
